+++ HISTORY.md
@@ -60,6 +60,7 @@
 ### 2019-04-05
 * Objects: eliminated "clicks" when playing music notes
 * Objects: eliminated "clicks" when playing a frequency
+* Widgets, Objects: Adjusted PianoKeyboard for the new audio engine
 
 ### 2019-04-04
 * Objects, Threads: new "play frequency" commands in the Sounds category